chore: move GatewaySpecError to gateway_types
commit-id:ecf04912
