+++ crates/gateway/src/compilation.rs
@@ -4,6 +4,7 @@ use cairo_lang_starknet_classes::casm_contract_class::CasmContractClass;
 use cairo_lang_starknet_classes::contract_class::ContractClass as CairoLangContractClass;
 use starknet_api::contract_class::ClassInfo;
 use starknet_api::rpc_transaction::RpcDeclareTransaction;
+use starknet_gateway_types::errors::GatewaySpecError;
 use starknet_sierra_compile::cairo_lang_compiler::CairoLangSierraToCasmCompiler;
 use starknet_sierra_compile::command_line_compiler::CommandLineCompiler;
 use starknet_sierra_compile::config::SierraToCasmCompilationConfig;
@@ -11,7 +12,7 @@ use starknet_sierra_compile::utils::into_contract_class_for_compilation;
 use starknet_sierra_compile::SierraToCasmCompiler;
 use tracing::{debug, error};
 
-use crate::errors::{GatewayResult, GatewaySpecError};
+use crate::errors::GatewayResult;
 
 #[cfg(test)]
 #[path = "compilation_test.rs"]
