+++ crates/http_server/src/http_server.rs
@@ -8,8 +8,9 @@ use axum::{Json, Router};
 use starknet_api::rpc_transaction::RpcTransaction;
 use starknet_api::transaction::TransactionHash;
 use starknet_gateway::config::{GatewayConfig, GatewayNetworkConfig};
-use starknet_gateway::errors::{GatewayRunError, GatewaySpecError};
+use starknet_gateway::errors::GatewayRunError;
 use starknet_gateway_types::communication::SharedGatewayClient;
+use starknet_gateway_types::errors::GatewaySpecError;
 use starknet_gateway_types::gateway_types::{GatewayInput, MessageMetadata};
 use starknet_mempool_infra::component_runner::{ComponentStartError, ComponentStarter};
 use tracing::{error, info, instrument};
