+++ crates/tests-integration/Cargo.toml
@@ -24,6 +24,7 @@ starknet-types-core.workspace = true
 starknet_api.workspace = true
 starknet_client.workspace = true
 starknet_gateway = { workspace = true, features = ["testing"] }
+starknet_gateway_types.workspace = true
 starknet_mempool_infra.workspace = true
 starknet_mempool_node.workspace = true
 starknet_mempool_types.workspace = true
