+++ crates/tests-integration/src/integration_test_setup.rs
@@ -6,7 +6,7 @@ use starknet_api::executable_transaction::Transaction;
 use starknet_api::rpc_transaction::RpcTransaction;
 use starknet_api::transaction::TransactionHash;
 use starknet_gateway::config::GatewayNetworkConfig;
-use starknet_gateway::errors::GatewaySpecError;
+use starknet_gateway_types::errors::GatewaySpecError;
 use starknet_mempool_infra::trace_util::configure_tracing;
 use starknet_mempool_node::servers::get_server_future;
 use starknet_mempool_node::utils::create_clients_servers_from_config;
